+++ semcore/data-table/src/index.ts
@@ -19,5 +19,19 @@ const wrapDataTable = <PropsExtending extends {}>(
   ) => React.ReactNode,
 ): DataTableType => wrapper as any;
 
-export { DataTable, ACCORDION, ROW_GROUP, UNIQ_ROW_KEY, wrapDataTable };
-export type { DataTableSort, DataTableData, DataTableProps, DataTableChangeSort };
+export {
+  DataTable,
+  ACCORDION,
+  ROW_GROUP,
+  /**
+   * @deprecated use property `uniqueRowKey` in DataTableProps to set key of unique value in your data.
+   */
+  UNIQ_ROW_KEY,
+  wrapDataTable,
+};
+export type {
+  DataTableSort,
+  DataTableData,
+  DataTableProps,
+  DataTableChangeSort,
+};